# specs/002-task-api-endpoints/contracts/openapi.yaml
openapi: 3.0.0
info:
  title: Todo App API - Task Endpoints
  version: 1.0.0
paths:
  /api/{user_id}/tasks:
    get:
      summary: List all tasks for a user
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: user_id
          schema:
            type: string
            format: uuid
          required: true
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskRead'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (Accessing another user's tasks)

    post:
      summary: Create a new task
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: user_id
          schema:
            type: string
            format: uuid
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: Task created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskRead'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden

  /api/{user_id}/tasks/{task_id}:
    get:
      summary: Get task details
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: user_id
          schema:
            type: string
            format: uuid
          required: true
        - in: path
          name: task_id
          schema:
            type: string
            format: uuid
          required: true
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskRead'
        '404':
          description: Task not found

    put:
      summary: Update task
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: user_id
          schema:
            type: string
            format: uuid
          required: true
        - in: path
          name: task_id
          schema:
            type: string
            format: uuid
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskRead'

    delete:
      summary: Delete task
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: user_id
          schema:
            type: string
            format: uuid
          required: true
        - in: path
          name: task_id
          schema:
            type: string
            format: uuid
          required: true
      responses:
        '204':
          description: Task deleted

  /api/{user_id}/tasks/{task_id}/complete:
    patch:
      summary: Toggle task completion
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: user_id
          schema:
            type: string
            format: uuid
          required: true
        - in: path
          name: task_id
          schema:
            type: string
            format: uuid
          required: true
      responses:
        '200':
          description: Task completion toggled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskRead'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    TaskCreate:
      type: object
      required:
        - content
      properties:
        content:
          type: string
          minLength: 1
    TaskUpdate:
      type: object
      properties:
        content:
          type: string
          minLength: 1
        completed:
          type: boolean
    TaskRead:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        content:
          type: string
        completed:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time